// src/Framework/SPPC.Tadbir.Web/ClientApp/app/model/viewRowPermission.ts
// ------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated by a tool.
//     Runtime Version: 1.1.349
//     Template Version: 1.0
//     Generation Date: 07/22/2018 12:50:26 ب.ظ
//  
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.
// </auto-generated>
// ------------------------------------------------------------------------------


export interface ViewRowPermission {
    roleId: number;
    viewId: number;
    id: number;
    accessMode: string;
    value: number;
    value2: number;
    textValue: string;
    items: number[];
}